Clean up schema errors
