+++ pg_lakehouse/src/hooks/mod.rs
@@ -16,12 +16,10 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 mod executor;
-mod explain;
 mod query;
 
 use async_std::task::block_on;
 use pgrx::*;
-use std::ffi::CStr;
 
 pub struct LakehouseHook;
