+++ pg_lakehouse/src/schema/mod.rs
@@ -15,6 +15,5 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-pub mod attribute;
 pub mod cell;
 pub mod datetime;
